requirements.txt
Django==3.2.4
django-filebrowser-no-grappelli2==3.8.4
-e git+git://github.com/christianwgd/django-filebrowser-no-grappelli.git#egg=django-filebrowser-no-grappelli
django-tinymce==3.3.0
Pillow==8.2.0